# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Test task API Documentation",
        "description": "Test task API Documentation Description",
        "contact": {
            "name": "Swagger API Team"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://task.firscoders.shop/api/",
            "description": "Global API server"
        },
        {
            "url": "http://test-task.loc/api/",
            "description": "Local API server"
        }
    ],
    "paths": {
...
        "/v1/products": {
            "get": {
                "tags": [
                    "Products Routes"
                ],
                "summary": "Products list",
                "description": "Products list.",
                "operationId": "productsList",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    },
                    {
                        "name": "perpage",
                        "in": "query",
                        "description": "Per page number",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "15"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Products Routes"
                ],
                "summary": "Product add",
                "description": "Product add.",
                "operationId": "productCreate",
                "parameters": [
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Product name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Product Name"
                    },
                    {
                        "name": "amount",
                        "in": "query",
                        "description": "Product price",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "example": "130000"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/products/{product_id}": {
            "get": {
                "tags": [
                    "Products Routes"
                ],
                "summary": "Product show",
                "description": "Product show.",
                "operationId": "productShow",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "2"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Products Routes"
                ],
                "summary": "Product update",
                "description": "Product update.",
                "operationId": "productUpdate",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "143"
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "Product title",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "New Product Name"
                    },
                    {
                        "name": "amount",
                        "in": "query",
                        "description": "New Product price",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "132000"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Products Routes"
                ],
                "summary": "Product delete",
                "description": "Product edit info.",
                "operationId": "productDelete",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "143"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Delete error!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/v1/products/{product_id}/edit": {
            "get": {
                "tags": [
                    "Products Routes"
                ],
                "summary": "Product edit info",
                "description": "Product edit info.",
                "operationId": "productEdit",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "Product id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "2"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success!",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}